+++ CHANGELOG.md
@@ -7,13 +7,14 @@
 ### Feature
 
 - Include a pluggable architecture for pluggable render-time insertions (similar to <Portal>) @tiberiuichim
+
 ### Bugfix
 
 - Include selected block in multiselections @sneridagh
+- In BlockDataForm, always clone schema before applying enhancers @tiberiuichim
+- In BlockDataForm, don't add the variations field multiple times @tiberiuichim
 
 ### Internal
-
-- Upgrade Storybook to 6.2.2 @tiberiuichim
 
 - Implement Github actions workflow to deploy the documentation to the Plone Foundation server @ericof
 - Pin `immutable` to an updated version that does not produce continuous deprecation notices in console on every change @sneridagh